+++ admin_redactions.php
@@ -51,6 +51,7 @@ require_admin_page();
     </div>
   </main>
 
+<script src="/js/date_format.js"></script>
 <script src="/js/admin_redactions.js"></script>
 </body>
 </html>